requirements.txt
langchain
langchain-community
langchain_groq
pypdf
faiss-cpu
sentence-transformers
groq
python-dotenv
requests
streamlit